requirements.txt
# --index-url https://download.pytorch.org/whl/cu118

tqdm
rotary_embedding_torch
transformers==4.31.0
tokenizers==0.13.3
inflect
progressbar
einops==0.4.1
unidecode
scipy
librosa==0.9.1
ffmpeg
numpy
numba
# torch==2.0.1+cu118
# torchaudio==2.0.2
torchaudio
threadpoolctl
llvmlite
appdirs
nbconvert==5.3.1
pydantic==1.9.1
deepspeed==0.8.3
py-cpuinfo
hjson
psutil
bentoml>=1.0.22

